pyproject.toml
[tool.poetry]
name = "incolume.py.vsaliens"
version = "0.2.0-alpha.1"
description = ""
authors = ["Britodfbr <britodfbr@gmail.com>"]
packages = [
  {include = "incolume/py", from=""},
]

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.1.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
